% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{adj_to_graph}
\alias{adj_to_graph}
\title{Convert a adjacency matrix to graph.}
\usage{
adj_to_graph(adj_matrix, mode)
}
\description{
This is used mainly for converting the estimated matrix to a graph.
Using the lower triangular matrix corresponds to following the factorisation
in the natural order, i.e. from the first column to the last column of the data.
}
\keyword{internal}
