% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{analyze_variable}
\alias{analyze_variable}
\title{Detect variable type and decide what family of distribution to use}
\usage{
analyze_variable(x0, threshold = 5)
}
\arguments{
\item{x0;}{data vector}
}
\description{
Detect variable type and decide what family of distribution to use
}
\keyword{internal}
