% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CE_independence_test.R
\name{column_to_prob}
\alias{column_to_prob}
\title{Find the empirical multinomial distribution from a vector}
\usage{
column_to_prob(vec0, s)
}
\arguments{
\item{s}{number of classes}
}
\description{
Find the empirical multinomial distribution from a vector
}
\keyword{internal}
