% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{compare_graphs}
\alias{compare_graphs}
\title{Compare two graphs}
\usage{
compare_graphs(m1, m2, mode = "undirected", color_vec = c("green", "black",
  "red"))
}
\arguments{
\item{m1}{adjacency matrix}

\item{m2}{adjacency matrix}

\item{mode}{If "directed", the function uses only the upper triangular part of the
adjacency matrix; otherwise, the function uses the full matrix.}

\item{color_vec}{characters vector; see description for details.}
}
\description{
This function compares two graphs and labels the matched / unmatched
edges with different colors.
}
\details{
By default, green color applies to edges that are present in both graphs;
black color applies to edges that are present in the first graph but not the second
graph; red color applies to edges that are present in the second graph but
not the first graph.
}
