% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{compute_datum_likelihood}
\alias{compute_datum_likelihood}
\title{Compute the likelihood for a single data-point given the complete
factorisation table}
\usage{
compute_datum_likelihood(table0, x, log = TRUE)
}
\arguments{
\item{table0}{dataframe; A factorisation table, output from "build_conditional".}

\item{x}{data vector; one data point of multiple features/covariates.}

\item{log}{TRUE / FALSE; returns loglikelihood / likelihood.}
}
\description{
Compute the likelihood for a single data-point given the complete
factorisation table
}
\keyword{internal}
