% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{compute_likelihood}
\alias{compute_likelihood}
\title{Compute the likelihood for a single data-point given the complete
factorisation table}
\usage{
compute_likelihood(table0, data0, log = TRUE)
}
\arguments{
\item{table0}{dataframe; A factorisation table, output from "build_conditional".}

\item{data0}{matrix or dataframe, the data.}

\item{log}{TRUE / FALSE; returns loglikelihood / likelihood.}
}
\description{
Compute the likelihood for a single data-point given the complete
factorisation table
}
\keyword{internal}
