% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_cor.R
\name{copula_cor}
\alias{copula_cor}
\title{Copula correlation function}
\usage{
copula_cor(data0)
}
\description{
Copula correlation function
}
\references{
Copula Correlation: An Equitable Dependence Measure and
Extension of Pearson’s Correlation. https://arxiv.org/pdf/1312.7214.pdf

https://github.com/felixleungsc/copulr
}
\keyword{internal}
