% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{create_missingness}
\alias{create_missingness}
\title{Randomly remove entries from a dataset}
\usage{
create_missingness(data0, p = 0.1)
}
\arguments{
\item{data0}{data.frame or matrix; the data}

\item{p}{probability of missing; for each entry in the data, the
missingness follows a Bernoulli(p) distribution.}
}
\description{
Randomly remove entries from a dataset
}
