% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{create_random_edges}
\alias{create_random_edges}
\title{This function creates random number of edges}
\usage{
create_random_edges(num_nodes, zero_pos, p = 0.5)
}
\arguments{
\item{num_nodes}{Integer; number of nodes available to be connected.}

\item{zero_pos}{The position of the node that shouldn't have an edge.}

\item{p}{Probability of having an edge. This is a parameter to make sure
the desired number of edges is generated.}
}
\description{
This function creates random number of edges
}
\keyword{internal}
