% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{create_random_graph}
\alias{create_random_graph}
\title{This function generates a random graph in adjacency matrix notation.}
\usage{
create_random_graph(num_nodes, p = 0.5)
}
\arguments{
\item{num_nodes}{Integer; number of nodes available to be connected..}

\item{p}{Between 0 and 1; additional parameter to control number of edges to be generated. See details.}
}
\description{
This function generates a random graph in adjacency matrix notation.
}
\details{
An edge is generated using Bernoulli(p) distribution.
}
