% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{dispersion2parameters}
\alias{dispersion2parameters}
\title{Convert dispersion value into parameter value}
\usage{
dispersion2parameters(parameters, dispersion, family)
}
\description{
Convert dispersion value into parameter value
}
\keyword{internal}
