% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorisation.R
\name{factorise}
\alias{factorise}
\title{The function generates a factorisation based on a graph}
\usage{
factorise(rgraph, response_node = 1)
}
\arguments{
\item{rgraph}{matrix; graph represented by an adjacency matrix}

\item{response_node}{the index of the response variable; simply ignore if there is
no response variable.}
}
\description{
The function generates a factorisation based on a graph
}
