% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorisation.R
\name{factorise_density}
\alias{factorise_density}
\title{Factorise the joint density of k+1 variables into
conditional x joint density of k variables}
\usage{
factorise_density(joint_vec, keep_pos = 1)
}
\arguments{
\item{joint_vec}{The indexes of variables in a joint density}

\item{keep_index}{The position of the variable to be kept}
}
\description{
Factorise the joint density of k+1 variables into
conditional x joint density of k variables
}
\keyword{internal}
