% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{family2invLinkFUN}
\alias{family2invLinkFUN}
\title{Return the inverse link function of a family object}
\usage{
family2invLinkFUN(family, ...)
}
\description{
Return the inverse link function of a family object
}
\keyword{internal}
