% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{family2likeFUN}
\alias{family2likeFUN}
\title{Return the density function of a family object}
\usage{
family2likeFUN(family, ...)
}
\description{
Return the density function of a family object
}
\keyword{internal}
