% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{family2parameters}
\alias{family2parameters}
\title{Return the parameter of a family object}
\usage{
family2parameters(family, ...)
}
\description{
Return the parameter of a family object
}
\keyword{internal}
