% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{family2simFUN}
\alias{family2simFUN}
\title{Return the sampling function of a family object}
\usage{
family2simFUN(family, ...)
}
\description{
Return the sampling function of a family object
}
\keyword{internal}
