% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_density.R
\name{fast_build_conditional}
\alias{fast_build_conditional}
\title{Optimise the "build_conditional" function for gibbs sampler model selection.}
\usage{
fast_build_conditional(df0, family_df0)
}
\description{
Optimise the "build_conditional" function for gibbs sampler model selection.
}
\keyword{internal}
