% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{fit_glm}
\alias{fit_glm}
\title{# Helper functions
#' @keywords internal
get_nll_from_aic <- function(aic_score, num_var) {
  (aic_score - 2*num_var) / (-2)
}}
\usage{
fit_glm(x, y, family, engine)
}
\description{
# Helper functions
#' @keywords internal
get_nll_from_aic <- function(aic_score, num_var) {
  (aic_score - 2*num_var) / (-2)
}
}
\keyword{internal}
