% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_helper.R
\name{fit_graph}
\alias{fit_graph}
\title{Wrapper function for 'factorise', 'build_conditional' and 'MLE_graph'}
\usage{
fit_graph(rgraph, family, data0)
}
\arguments{
\item{rgraph}{matrix; a graph encoded in adjacency matrix.}

\item{family}{vector of characters string; the family for each variable.}

\item{data0}{dataframe; the data.}
}
\description{
This function takes a graph and fits the data with the family
specification.
}
