% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{get_model_likelihood}
\alias{get_model_likelihood}
\title{Get likelihood from a fitted model}
\usage{
get_model_likelihood(table0)
}
\arguments{
\item{table0}{data.frame; the fitted graph from 'MLE_graph'.}
}
\value{
The loglikelihood value.
}
\description{
Get likelihood from a fitted model
}
