% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmGraph.R
\docType{package}
\name{glmGraph}
\alias{glmGraph}
\alias{glmGraph-package}
\title{A package for simulating data from a GLM model}
\description{
This packages provides functions to simulate from and fit graphical models
where all conditional distributions follow the generalised linear model. This is
mainly used to construct an approximation to a multivariate distribution.
}
\author{
Jackson Kwok, Felix leung
}
