% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{imputation}
\alias{imputation}
\title{Perform imputation based on a graphical model}
\usage{
imputation(table0, data0, method = "response", threshold = 0.5)
}
\arguments{
\item{table0}{data.frame; full joint density specification fitted to data.}

\item{data0}{data.frame; the data.}

\item{method}{"mean", "response" or "sample". See 'details' for more information.}

\item{threshold}{probability threshold to decide the class for the binomial distribution.}
}
\description{
Perform imputation based on a graphical model
}
\details{
"mean" refers to imputation with the conditional mean.
"response" refers to imputation with the fitted class / value.
"sample" refers to imputation with a sample from the conditional distribution.
"sample" is commonly used to do multiple imputation.
Note that "mean" and "response" are the same for continuous distributions.
}
