% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph.R
\name{is_DAG}
\alias{is_DAG}
\title{This function checks if an adjacency matrix is a DAG.}
\usage{
is_DAG(m0)
}
\arguments{
\item{m0}{Numeric matrix; an adjacency matrix.}
}
\description{
This function checks if an adjacency matrix is a DAG.
}
