% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_fitted}
\alias{is_fitted}
\title{Check if the model has been fitted to some data.}
\usage{
is_fitted(df0)
}
\arguments{
\item{df0}{Data.frame; the graphical GLM.}
}
\description{
Check if the model has been fitted to some data.
}
