% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{learn_graph}
\alias{learn_graph}
\title{Model selection for GLM graph}
\usage{
learn_graph(data0, p = 0.2, init_rgraph, lambda, reg = "BIC",
  num_iter = 100, burn_ins = 10, ...)
}
\arguments{
\item{data0}{dataframe; the data.}

\item{p}{Between 0 and 1; prior brief of how connected the graph is. If init_rgraph is
specified, then this is not used.}

\item{init_rgraph}{adjacency matrix (optional); the graph to begin estimation with.}

\item{lambda}{Tuning parameter for gibbs sampler.}

\item{reg}{Information criteria for regularisation; "AIC" or "BIC".}

\item{num_iter}{number of iterations for gibbs sampler.}

\item{burn_ins}{number of burn-ins for gibbs sampler.}

\item{...}{Extra parameters to be passed to speedglm::speedglm.}
}
\description{
Model selection for GLM graph
}
