% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_entropy_MS.R
\name{learn_graph_by_CE}
\alias{learn_graph_by_CE}
\title{Model selection for GLM graphical model using cross-entropy method}
\usage{
learn_graph_by_CE(data0, rho = 0.1, batch_size = 50, tol = 1e-05,
  alpha = 1, graph_init = "random", threshold = T)
}
\arguments{
\item{data0}{dataframe; the data.}

\item{rho}{Between 0 and 1, samples with a "score" above the (1-rho)-quantile are
considered as good samples.}

\item{batch_size}{Number of samples to generate in each iteration.}

\item{tol}{Tolerence of the threshold for convergence}

\item{alpha}{Smooth-update parameter between 0 and 1. When alpha = 1, the update uses
only the current estimate; when alpha = 0.5, the update uses half of the current estimate
and half of the last estimate.}

\item{graph_init}{characters string; one of "random", "correlation" and "mutual".
"random" generates a graph randomly. "correlation" computes the pairwise correlation
between variables and keeps the ones above the third quartile. "mutual" is similar to
"correlation" except it uses pairwise mutual information instead.}

\item{threshold}{if TRUE, the function returns a graph; if FALSE, the function returns
the underlying Bernoulli probabilities.}
}
\description{
Model selection for GLM graphical model using cross-entropy method
}
