% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CE_independence_test.R
\name{learn_graph_by_independence_CE}
\alias{learn_graph_by_independence_CE}
\title{Learn graph structure using the independence-based approach}
\usage{
learn_graph_by_independence_CE(data0, method, batch_size = 500, rho = 0.1,
  tol = 1e-05, alpha = 1)
}
\arguments{
\item{data0}{dataframe; the data.}

\item{method}{One of 'random', 'correlation', 'mutual' and 'copula".}

\item{batch_size}{Number of samples to generate in each iteration.}

\item{rho}{Between 0 and 1, samples with a "score" above the (1-rho)-quantile are
considered as good samples.}

\item{tol}{Tolerence of the threshold for convergence}

\item{alpha}{Smooth-update parameter between 0 and 1. When alpha = 1, the update uses
only the current estimate; when alpha = 0.5, the update uses half of the current estimate
and half of the last estimate.}
}
\description{
Learn graph structure using the independence-based approach
}
\keyword{internal}
