% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_independence_test.R
\name{learn_graph_by_independence_Gibbs}
\alias{learn_graph_by_independence_Gibbs}
\title{Learn graph structure using the independence-based approach}
\usage{
learn_graph_by_independence_Gibbs(data0, method, lambda, num_iter = 100,
  reg_FUN = "BIC")
}
\arguments{
\item{data0}{dataframe; the data.}

\item{method}{One of 'random', 'correlation', 'mutual' and 'copula".}

\item{lambda}{Tuning parameter for gibbs sampler.}

\item{num_iter}{number of iterations for gibbs sampler.}
}
\description{
Learn graph structure using the independence-based approach
}
\keyword{internal}
