% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{plot_DAG}
\alias{plot_DAG}
\title{Plot a DAG}
\usage{
plot_DAG(adj_mat)
}
\arguments{
\item{adj_mat}{An adjacency matrix with columns named after the variables.}
}
\description{
Plot a DAG
}
\examples{
\dontrun{
m0 <- random_DAG(num_nodes = 6)
plot_DAG(m0)
}
}
