% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{plot_graph}
\alias{plot_graph}
\title{This function plots graph from the function "create_random_graph".}
\usage{
plot_graph(rgraph, vertex.size = 30, ...)
}
\arguments{
\item{rgraph}{matrix; output from "create_random_graph".}
}
\description{
This function plots graph from the function "create_random_graph".
}
