% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_and_prediction.R
\name{predict.gglm.data.frame}
\alias{predict.gglm.data.frame}
\title{Predict based on a graphical model}
\usage{
\method{predict}{gglm.data.frame}(object, data0, resp_var, method = "mean",
  threshold = 0.5, ...)
}
\arguments{
\item{object}{data.frame; full joint density specification fitted to data.}

\item{data0}{data.frame; the data.}

\item{resp_var}{character strings; the name of the variable to be predicted.}

\item{method}{"mean" or "response". See 'details' for more information.}

\item{threshold}{probability threshold to decide the class for the binomial distribution.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Predict based on a graphical model
}
\details{
"mean" refers to predicting with the conditional mean.
A common usage is to predict the probability for the occurrence of a class.
"response" refers to predicting with the fitted class / value.
A common usage is to predict the actual class.
Note that "mean" and "response" are the same for continuous distribution.
}
