% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.graph}
\alias{predict.graph}
\title{Predict based on a graphical model}
\usage{
predict.graph(table0, data0, method = "mean", threshold = 0.5)
}
\arguments{
\item{table0}{data.frame; full joint density specification fitted to data.}

\item{data0}{data.frame; the data.}

\item{method}{"mean", "response" or "sample". See 'details' for more information.}

\item{threshold}{probability threshold to decide the class for the binomial distribution.}
}
\description{
Predict based on a graphical model
}
\details{
"mean" refers to predicting with the conditional mean.
"response" refers to predicting with the fitted class / value.
"sample" refers to predicting with a sample from the conditional distribution.
A common usage for "mean" is to predict the probability for the occurrence of a class,
while a common usage for "response" is to predict the actual class.
Note that "mean" and "response" are the same for continuous distribution.
}
