% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_and_prediction.R
\name{predict_graph}
\alias{predict_graph}
\title{Predict based on a graphical model}
\usage{
predict_graph(table0, data0, method = "mean", threshold = 0.5)
}
\description{
Predict based on a graphical model
}
\keyword{internal}
