% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{print.gglm.data.frame}
\alias{print.gglm.data.frame}
\title{Print out summary of a complete factorisation table}
\usage{
\method{print}{gglm.data.frame}(x, ...)
}
\arguments{
\item{x}{dataframe; the complete factorisation table; output from
build_conditional.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print out summary of a complete factorisation table
}
