% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{print_summary}
\alias{print_summary}
\title{Print out summary of a complete factorisation table}
\usage{
print_summary(table0)
}
\arguments{
\item{table0}{dataframe; the complete factorisation table; output from
build_conditional.}
}
\description{
Print out summary of a complete factorisation table
}
