% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{random_DAG}
\alias{random_DAG}
\title{This function generates a random directed acyclic graph (DAG) in adjacency matrix notation.}
\usage{
random_DAG(num_nodes, p = 0.5)
}
\arguments{
\item{num_nodes}{Integer; number of nodes available to be connected..}

\item{p}{p Between 0 and 1; a parameter to control number of edges to be generated.
The higher this number is, the more connected the graph will be.}
}
\description{
This function generates a random directed acyclic graph (DAG) in adjacency matrix notation.
}
