% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_entropy_MS.R
\name{sample_random_matrix}
\alias{sample_random_matrix}
\title{Sample based on a matrix of probabilities}
\usage{
sample_random_matrix(matrix0, prob_vec)
}
\arguments{
\item{matrix0}{empty_matrix}

\item{prob_vec}{Should have length (d * (d-1) / 2)}
}
\description{
Sample based on a matrix of probabilities
}
\keyword{internal}
