% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CE_independence_test.R
\name{sample_rank_threshold}
\alias{sample_rank_threshold}
\title{Sample a threshold for the rank}
\usage{
sample_rank_threshold(list0)
}
\description{
Sample a threshold for the rank
}
\keyword{internal}
