% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{select_graph}
\alias{select_graph}
\title{Model selection for GLM graph}
\usage{
select_graph(data0, p = 0.2, lambda = 1, num_iter = 100,
  graph_init = "random")
}
\arguments{
\item{data0}{dataframe; the data.}

\item{p}{Between 0 and 1; prior brief of how connected the graph is.}

\item{lambda}{Tuning parameter for gibbs sampler.}

\item{num_iter}{number of iterations for gibbs sampler.}

\item{graph_init}{characters string; one of "random", "correlation" and "mutual".
"random" generates a graph randomly. "correlation" computes the pairwise correlation
between variables and keeps the ones above the third quartile. "mutual" is similar to
"correlation" except it uses pairwise mutual information instead.}
}
\description{
Model selection for GLM graph
}
