% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorisation.R
\name{simplify_conditional}
\alias{simplify_conditional}
\title{This function simplifies conditionanl given the dependence information}
\usage{
simplify_conditional(conditional_list, edges)
}
\arguments{
\item{conditional_list}{The conditional density. It should be a list of two vectors.}

\item{edges}{The vector containing the dependence information.}
}
\description{
This function simplifies conditionanl given the dependence information
}
