% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_datum}
\alias{simulate_datum}
\title{Simulate a datapoint given the complete factorisation table}
\usage{
simulate_datum(table0)
}
\description{
Simulate a datapoint given the complete factorisation table
}
\keyword{internal}
