% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_density.R
\name{update_conditional_mean}
\alias{update_conditional_mean}
\title{Update the conditional mean given the data}
\usage{
update_conditional_mean(df0, x0)
}
\description{
Update the conditional mean given the data
}
\section{Functions}{
\itemize{
\item \code{update_conditional_mean}: function updates the conditional mean given the data in preparation
for the parameters update.
}}

\keyword{internal}
