% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CE_independence_test.R
\name{threshold_rank_matrix}
\alias{threshold_rank_matrix}
\title{Create a graph by applying the threshold to the rank matrix}
\usage{
threshold_rank_matrix(m0, threshold0)
}
\description{
Create a graph by applying the threshold to the rank matrix
}
\keyword{internal}
