% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CE_independence_test.R
\name{update_multinom_prob}
\alias{update_multinom_prob}
\title{Take the threshold samples, marginalise them and turn them into multinomial
probabilities. One row is one sample.}
\usage{
update_multinom_prob(prob_list0, t_matrix, alpha)
}
\description{
Take the threshold samples, marginalise them and turn them into multinomial
probabilities. One row is one sample.
}
\keyword{internal}
