% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_density.R
\name{build_conditional}
\alias{build_conditional}
\title{Specify the distribution for each conditional density}
\usage{
build_conditional(df0, family)
}
\arguments{
\item{df0}{A factorisation table; output from "factorise".}

\item{family}{Vector of characters string; names of the exponential family of
distribution to be used. It must be one of c("gaussian", "gamma", "poisson",
"binomial")}
}
\description{
Specify the distribution for each conditional density
}
