% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{compute_likelihood}
\alias{compute_likelihood}
\title{Compute the likelihood for a dataset given the complete factorisation table}
\usage{
compute_likelihood(table0, data0, log = TRUE)
}
\arguments{
\item{table0}{dataframe; A factorisation table, output from "build_conditional".}

\item{data0}{matrix or dataframe, the data.}

\item{log}{TRUE / FALSE; returns loglikelihood / likelihood.}
}
\description{
Compute the likelihood for a dataset given the complete factorisation table
}
\note{
This function computes the loglikelihood directly from the data, rather than
from the AIC values of the fitted model, so that could be slight rounding difference
to the likelihood extracted from the fitted model. One should expect the difference
to be less than the number of nodes.
}
