% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{factorise}
\alias{factorise}
\title{The function generates a factorisation of a joint density based on a graph.}
\usage{
factorise(rgraph)
}
\arguments{
\item{rgraph}{matrix; graph represented by an adjacency matrix}
}
\description{
The function generates a factorisation of a joint density based on a graph.
}
\examples{
m0 <- random_DAG(5)
factorise(m0)
}
