% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorisation.R
\name{factorise}
\alias{factorise}
\title{The function generates a factorisation based on a graph}
\usage{
factorise(rgraph)
}
\arguments{
\item{rgraph}{matrix; graph represented by an adjacency matrix}
}
\description{
The function generates a factorisation based on a graph
}
