% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{imputation}
\alias{imputation}
\title{Perform imputation based on a graphical model}
\usage{
imputation(table0, data0)
}
\arguments{
\item{table0}{data.frame; full joint density specification fitted to data}
}
\description{
Perform imputation based on a graphical model
}
