% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{learn_graph}
\alias{learn_graph}
\title{Model selection for GLM graph}
\usage{
learn_graph(data0, p = 0.2, lambda, num_iter = 100, burn_ins = 10,
  graph_init = "random", reg = "BIC", ...)
}
\arguments{
\item{data0}{dataframe; the data.}

\item{p}{Between 0 and 1; prior brief of how connected the graph is.}

\item{lambda}{Tuning parameter for gibbs sampler.}

\item{num_iter}{number of iterations for gibbs sampler.}

\item{burn_ins}{number of burn-ins for gibbs sampler.}

\item{graph_init}{characters string; one of "random", "correlation" and "mutual".
"random" generates a graph randomly. "correlation" computes the pairwise correlation
between variables and keeps the ones above the third quartile. "mutual" is similar to
"correlation" except it uses pairwise mutual information instead.}

\item{reg}{Information criteria for regularisation; "AIC" or "BIC".}

\item{...}{Extra parameters to be passed to speedglm::speedglm.}
}
\description{
Model selection for GLM graph
}
