% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_independence_test.R
\name{learn_graph_by_independence_CD}
\alias{learn_graph_by_independence_CD}
\title{Learn graph structure using the independence-based approach (coordinate descent)}
\usage{
learn_graph_by_independence_CD(data0, method, reg = "BIC")
}
\arguments{
\item{data0}{dataframe; the data.}

\item{method}{One of 'random', 'correlation', 'mutual' and 'copula".}

\item{reg}{Information criteria for regularisation; "AIC" or "BIC".}
}
\description{
Learn graph structure using the independence-based approach (coordinate descent)
}
