% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{mean2parameters}
\alias{mean2parameters}
\title{Convert (conditonal) mean value into parameter value}
\usage{
mean2parameters(parameters, mu, family)
}
\arguments{
\item{parameters}{a named list}

\item{mu}{numeric; the mean value}

\item{family0}{characters string; the exponential family.}
}
\description{
Convert (conditonal) mean value into parameter value
}
\keyword{internal}
