% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{MLE_graph}
\alias{MLE_graph}
\title{Fitting the GLM graphical model with MLE.}
\usage{
MLE_graph(table0, data0, engine = glm.fit)
}
\arguments{
\item{table0}{dataframe; the complete factorisation table, output from "build_conditional".}

\item{data0}{dataframe; the data.}

\item{engine}{function; the function to fit a GLM.}
}
\description{
Fitting the GLM graphical model with MLE.
}
