% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{select_graph}
\alias{select_graph}
\title{Model selection for GLM graph}
\usage{
select_graph(data0, p = 0.2, lambda = 1, num_iter = 100)
}
\arguments{
\item{data0}{dataframe; the data.}

\item{p}{Between 0 and 1; prior brief of how connected the graph is.}

\item{lambda}{Tuning parameter for gibbs sampler.}

\item{num_iter}{number of iterations for gibbs sampler.}
}
\description{
Model selection for GLM graph
}
